Game Loop template is created.
In order for this to run on your machines, you will need to resolve the Include directories and Linker files/directories in your project settings. They need to point at files or directories in the folder called "SDL2-2.0.10. Watch this short video to get the poject settings right:

https://www.youtube.com/watch?v=QQzAHcojEKg

...or watch this one if you are on a Mac:

https://www.youtube.com/watch?v=lHAUc6F_B3Y

The tutorial I used to create this is here:

https://www.youtube.com/watch?v=44tO977slsU

One file was not found during compile (maybe), but it could also just be a bug. I got around it by placing '''#define SDL_MAIN_HANDLED''' at the top of my Main.cpp file. It will probably resolve itself as the program is further developed.