Project header, cpp and asset files up to the end of #12
This is a gutted local directory based on the tutorial series "Let's Make Games" on Youtube, with minor changes and my own test assets. The files that are "missing" here are Visual Studio 2017 mumbo-jumbo (which, by the way, is extremely important but I can't upload it all on GitHub because the size of the directory is blowing up since I am copying the same directory 12 times and building from where I leave off. I do not know exactly why this happens, but I am guessing there are duplicated refrerences somewhere that need to be cleaned on my end. You will not have this problem if you dump these files into your own empty project.