This is my project directory iff you have Windows and VS2017
"ECS" := "Entity-Component System". This is very similar to how Unity works. An "Entity" is an object that can have a a component attached to it. These components have override methods for init() (initialize), update(), draw(), and/or render(). This keeps us from having to have a laundry list of update/render/draw methods in the Game.cpp (game engine). If we know what we want a component to do, we say what that is during class definition.