last video in progress 22/22. Pushing before major changes
I normalized the character's velocity vector.

I am working on shooting animations in prep for projectiles. In order to have richer animations, I need the program to believe the source image size is 96x96 instead of 32x32. Doing this is going to require changing a lot of inputs and carefully changing the character's collision region. The file sprite sheet's file size is going to increase by (almost) a factor of 9, but there are some redundant sprites I can eliminate now from the image now. These changes will be in the next update, hopefully.