Resolved Issue 1: destroying monster no longer breaks game
Problem was that a *specific monster*'s collider was being passed into the collision function, instead of a pointer to members in a vector of monsters, which would be skipped if it were empty. Lightbulb turned on, easy fix. We still need to refactor animation to distinguish between finite and continuous animation loops, and allow input to interrupt an animation in progress.